Add Session.location_info()
